requirements.txt
openai>=1.0.0
rich
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dotenv